slang: Add support for generating a subset of IRBuilder methods
This change adds support for generating the getter IRBuilder methods
for the IR instructions defined in slang-ir-insts.lua with explicit operands.

Since there was a mismatch in the type of getMetalMeshType's topology parameter
compared to the type described in the Lua definition of MetalMeshType, the getter
for this type is updated to reflect IRIntLit (as defined in the Lua definition of the type)
instead of IRInst.

Clients of getMetalMeshType are updated to reflect the change in its signature.

The definition of getMetalMeshType is also moved to slang-ir.cpp since Fiddle
only generates the signature of the getters and not their bodies.

Closes #7185